backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend service for Ad Block for Podcasts project"
authors = ["sanderskofsrud <sandesko@stud.ntnu.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
pydub = "^0.25.1"
openai-whisper = "^20240930"
transformers = "^4.46.0"
soundfile = "^0.12.1"
azure-cognitiveservices-speech = "^1.41.1"
python-dotenv = "^1.0.1"
openai = "1.54.3"
azure-ai-textanalytics = "^5.3.0"
matplotlib = "^3.7.2"
seaborn = "^0.13.2"
mitmproxy = "^10.3.0"
editdistance = "^0.8.1"
jinja2 = "^3.0.3"
pandas = "2.2.3"
plotly = "^5.3.1"

# Specify CUDA-enabled torch packages via direct URLs
# Uncomment and adjust these if you need CUDA support
# torch = { url = "https://download.pytorch.org/whl/cu118/torch-2.5.1%2Bcu118-cp311-cp311-win_amd64.whl" }
# torchvision = { url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp311-cp311-win_amd64.whl" }
# torchaudio = { url = "https://download.pytorch.org/whl/cu118/torchaudio-2.5.1%2Bcu118-cp311-cp311-win_amd64.whl" }

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "podcast_processor"
